feat : add article components to development page
