+++ pages/index.page.tsx
@@ -1,46 +1,7 @@
-import type { NextPage, GetStaticProps, InferGetStaticPropsType } from 'next';
-import { Article, HeaderBar } from 'src/components/molecules';
-import { connectMongo } from 'pages/api/middlewares/connectMongo';
-import ArticleModel from 'pages/api/models/articleModel';
+import type { NextPage } from 'next';
 
-const Home: NextPage = ({
-  articles,
-}: InferGetStaticPropsType<typeof getStaticProps>) => {
-  return (
-    <>
-      {articles.map((article: any) => (
-        <Article
-          key={article._id}
-          path={encodeURI(article._id)}
-          title={article.title}
-          description={article.description}
-        />
-      ))}
-    </>
-  );
+const Home: NextPage = () => {
+  return <div>This is Main Page</div>;
 };
 
 export default Home;
-
-export const getStaticProps: GetStaticProps = async () => {
-  try {
-    console.log('CONNECTING TO MONGO IN MAIN');
-    await connectMongo();
-    console.log('CONNECTED TO MONGO IN MAIN');
-
-    console.log('FETCHING DATA');
-    const articles = await ArticleModel.find();
-    console.log('FETCHED DATA');
-
-    return {
-      props: {
-        articles: JSON.parse(JSON.stringify(articles)),
-      },
-    };
-  } catch (error) {
-    console.log(error);
-    return {
-      notFound: true,
-    };
-  }
-};
